pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "mloader"
version = "1.2.3"
description = "Command-line tool to download manga from mangaplus"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Hurlenko" },
]
dependencies = [
    "Click>=8.3.1",
    "protobuf>=6.33.5,<7",
    "requests>=2.32.5",
    "Pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]

[project.urls]
Source = "https://github.com/hurlenko/mloader"

[project.scripts]
mloader = "mloader.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.2.0",
    "ruff>=0.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mloader"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = ["mloader/response_pb2.py"]

[tool.ruff.lint]
select = [
    "F",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN201",
    "ANN202",
    "ANN204",
    "ANN205",
]

[tool.ruff.lint.per-file-ignores]
"mloader/__main__.py" = ["E402"]
"mloader/main.py" = ["E402"]

[tool.coverage.run]
source = ["mloader"]
omit = [
    "mloader/response_pb2.py",
    "mloader/**/__init__.py",
]

[tool.coverage.report]
show_missing = true

[tool.black]
target_version = ["py313"]
line-length = 100
